Compatibility changes for wpa_supplicant and OpenSSH.
OpenSSH, especially, does some terrible things that mean that it needs
the EVP_CIPHER structure to be exposed ☹. Damian is open to a better API
to replace this, but only if OpenSSL agree too. Either way, it won't be
happening soon.

Reviewed-on: https://boringssl-review.googlesource.com/4330

+++ include/openssl/aes.h
@@ -146,9 +146,9 @@ OPENSSL_EXPORT void AES_cfb128_encrypt(const uint8_t *in, uint8_t *out,
  * Android they'll have no definition. */
 
 OPENSSL_EXPORT int AES_wrap_key(AES_KEY *key, const uint8_t *iv, uint8_t *out,
-                                const uint8_t *in, size_t in_len);
+                                const uint8_t *in, unsigned in_len);
 OPENSSL_EXPORT int AES_unwrap_key(AES_KEY *key, const uint8_t *iv, uint8_t *out,
-                                  const uint8_t *in, size_t in_len);
+                                  const uint8_t *in, unsigned in_len);
 
 
 #if defined(__cplusplus)
